+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 ## Bug Fixes
 
 * Weapon Item sheet tabs for Settings and Notes are now displaying properly.
+* Fixed an instance where Armor that held values for Shields or Characteristics modifiers, but had those fields not checked as "has," were still applying those values to the Actor.
 
 ## 0.2.4